requirements/local.txt
-r base.txt

django-model-utils
# DRF-spectacular for api documentation
drf-spectacular

Werkzeug[watchdog]
ipdb

watchgod
Collectfast

# Testing
# ------------------------------------------------------------------------------
mypy
django-stubs
pytest
pytest-sugar
djangorestframework-stubs
vcrpy
pytest-xdist
bfet
freezegun

# Documentation
# ------------------------------------------------------------------------------
sphinx
sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8
flake8-isort
coverage
black
pylint-django
pylint-celery
pre-commit
autoflake

# Django
# ------------------------------------------------------------------------------
django-debug-toolbar
django-extensions
django-coverage-plugin
pytest-django

uvicorn[standard]==0.18.2  # https://github.com/encode/uvicorn
channels

django-pghistory
